+++ splurge_exceptions/decorators/__init__.py
@@ -1,3 +1,0 @@
-"""Decorators module."""
-
-__all__: list[str] = []