+++ splurge_exceptions/managers/__init__.py
@@ -1,3 +1,0 @@
-"""Exception managers module."""
-
-__all__: list[str] = []