+++ splurge_exceptions/wrappers/__init__.py
@@ -1,3 +1,0 @@
-"""Exception wrappers module."""
-
-__all__: list[str] = []